CAR PRICE PREDICTION WITH MACHINE LEARNING
The price of a car depends on a lot of factors like the
goodwill of the brand of the car,features of the car,
horsepower and the mileage it gives and many more. Car
price prediction is one of the major research areas in
machine learning. So if you want to learn how to train a
car price prediction model then this project is for you.